app/requirements.txt
gdown==5.2.0
torch==2.3.0
matplotlib==3.9.0
seaborn==0.13.0
numpy==1.24.2
Flask==3.0.3
Pillow==10.2.0
transformers==4.40.2
safetensors==0.4.3
urllib3==1.26.14
Werkzeug==3.0.3 #or Werkzeug==2.2.2
tensorflow==2.16.1
torchvision==0.18.0
einops
timm
medmnist
pandas
scikit-learn
scikit-image
fvcore
tqdm
fire
torchattacks
fastervit

